+++ src/primitive.ts
@@ -17,9 +17,14 @@ function isSymbol(s: any): s is symbol {
   return type === 'symbol' || (type === 'object' && s != null && getTag(s) === '[object Symbol]');
 }
 
+function isNil(n: any): n is null | undefined {
+  return n == null;
+}
+
 export {
   isBoolean,
   isString,
   isNumber,
   isSymbol,
+  isNil,
 };